Create CollectionListViewRow - fixed a bug that allowed users to create a blank named collection
